+++ src/drivers/storage/YiiCacheStorage.php
@@ -6,13 +6,20 @@
 namespace putyourlightson\blitz\drivers\storage;
 
 use Craft;
-use ErrorException;
 use putyourlightson\blitz\events\RefreshCacheEvent;
 use putyourlightson\blitz\models\SiteUriModel;
 use yii\caching\CacheInterface;
 use yii\db\Exception;
 
 /**
+ * Uses Craft’s cache storage method, especially useful for distributed systems.
+ * The cached values can be compressed if gzip is supported, to help reduce the
+ * memory required for storage. While Gzip is not as fast as other compression
+ * algorithms such as Snappy and LZO, it is more widely supported and accepted
+ * by most browsers. This allows us to return compressed values to the browser
+ * directly, provided it accepts gzip encoding.
+ * https://docs.redis.com/latest/ri/memory-optimizations/#compress-values
+ *
  * @property-read string|null $settingsHtml
  */
 class YiiCacheStorage extends BaseCacheStorage
@@ -23,9 +30,19 @@ class YiiCacheStorage extends BaseCacheStorage
     public const KEY_PREFIX = 'blitz';
 
     /**
+     * @const string
+     */
+    public const ENCODING = 'gzip';
+
+    /**
      * @var string
      */
     public string $cacheComponent = 'cache';
+
+    /**
+     * @var bool Whether cached values should be compressed using gzip.
+     */
+    public bool $compressCachedValues = false;
 
     /**
      * @var CacheInterface|null
@@ -59,29 +76,42 @@ class YiiCacheStorage extends BaseCacheStorage
             return '';
         }
 
-        $value = '';
+        $key = $this->_getKey($siteUri);
 
-        // Redis cache can throw an exception if the connection is broken
-        try {
-            $value = $this->_cache->get($this->_getKey($siteUri));
+        if ($this->canCompressCachedValues()) {
+            $key[] = self::ENCODING;
+            $value = $this->_getFromCache($key);
 
-            // Decompress the value, if gzip is supported
-            if (function_exists('gzdecode')) {
-                // Catch E_WARNING level errors on failure
-                try {
-                    // Assign only if the decoded value is not `false`
-                    $value = gzdecode($value) ?: $value;
-                }
-                /** @noinspection PhpRedundantCatchClauseInspection */
-                catch (ErrorException) {
-                }
+            if ($value) {
+                $value = gzdecode($value);
             }
-        }
-        /** @noinspection PhpRedundantCatchClauseInspection */
-        catch (Exception) {
+        } else {
+            $value = $this->_cache->get($key);
         }
 
         return $value ?: '';
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getWithEncoding(SiteUriModel $siteUri, array $encodings = []): array
+    {
+        if ($this->_cache === null) {
+            return [null, null];
+        }
+
+        if ($this->canCompressCachedValues() && in_array(self::ENCODING, $encodings)) {
+            $key = $this->_getKey($siteUri);
+            $key[] = self::ENCODING;
+            $value = $this->_getFromCache($key);
+
+            if ($value) {
+                return [$value, self::ENCODING];
+            }
+        }
+
+        return [$this->get($siteUri), null];
     }
 
     /**
@@ -93,12 +123,14 @@ class YiiCacheStorage extends BaseCacheStorage
             return;
         }
 
-        // Compress the value, if gzip is supported
-        if (function_exists('gzencode')) {
+        $key = $this->_getKey($siteUri);
+
+        if ($this->canCompressCachedValues()) {
+            $key[] = self::ENCODING;
             $value = gzencode($value);
         }
 
-        $this->_cache->set($this->_getKey($siteUri), $value, $duration);
+        $this->_cache->set($key, $value, $duration);
     }
 
     /**
@@ -156,7 +188,13 @@ class YiiCacheStorage extends BaseCacheStorage
     {
         return Craft::$app->getView()->renderTemplate('blitz/_drivers/storage/yii-cache/settings', [
             'driver' => $this,
+            'compressionSupported' => function_exists('gzencode'),
         ]);
+    }
+
+    public function canCompressCachedValues(): bool
+    {
+        return $this->compressCachedValues && function_exists('gzencode');
     }
 
     /**
@@ -167,5 +205,24 @@ class YiiCacheStorage extends BaseCacheStorage
         // Cast the site ID to an integer to avoid an incorrect key
         // https://github.com/putyourlightson/craft-blitz/issues/257
         return [self::KEY_PREFIX, (int)$siteUri->siteId, $siteUri->uri];
+    }
+
+    /**
+     * Returns a key’s value from the cache.
+     */
+    private function _getFromCache(array $key): string
+    {
+        // Redis cache can throw an exception if the connection is broken
+        try {
+            $value = $this->_cache->get($key);
+
+            if ($value) {
+                return $value;
+            }
+        } /** @noinspection PhpRedundantCatchClauseInspection */
+        catch (Exception) {
+        }
+
+        return '';
     }
 }